refactor(presentation/loading): move the loading component to your folder
